Swap frame_rate on VideoCombine to float.
As video info produces a float for it's loaded frame rate, and
non_integer frame_rates are reasonably obtainable by using
select_every_nth, frame_rate has been changed on VideoCombine so that
the loaded frame_rate of Video info may be wired.

+++ videohelpersuite/nodes.py
@@ -161,7 +161,7 @@ class VideoCombine:
             "required": {
                 "images": ("IMAGE",),
                 "frame_rate": (
-                    "INT",
+                    "FLOAT",
                     {"default": 8, "min": 1, "step": 1},
                 ),
                 "loop_count": ("INT", {"default": 0, "min": 0, "max": 100, "step": 1}),
